+++ R/convert2ggplot.r
@@ -31,7 +31,7 @@
 #' 
 #' data(Pupfish)
 #' fit <- lm.rrpp(coords ~ log(CS) + Sex*Pop, SS.type = "I", 
-#' data = Pupfish, print.progress = FALSE, iter = 0) 
+#' data = Pupfish, print.progress = FALSE) 
 #' 
 #' # Predictions (holding alternative effects constant)
 #' 
@@ -56,7 +56,7 @@
 #'                                "Forelimb", "Hindlimb")])
 #'Y <- as.matrix(Y)
 #'R <- lm.rrpp(Y ~ SVL, data = PlethMorph, 
-#'             iter = 0, print.progress = FALSE)$LM$residuals
+#'              print.progress = FALSE)$LM$residuals
 #'
 #' # PCA (on correlation matrix)
 #'