+++ NEWS.md
@@ -4,6 +4,8 @@
 *  Added arguments for restricted resampling to `lm.rrpp` and `predict.lm.rrpp`, and permutation of full model residuals (with restrictions) to lm.rrpp
 * Added new function, `measurement.error`.
 * Added new function, `lm.rrpp.ws`.
+* Added `verbose` arguments to most analytical functions to reduce dense results.
+* Added utility functions (`getANOVAStats`, `getPermInfo`, `getTerms`, and `getModels`) to accommodate `verbose` function choices, plus allow users to easily extract objects as output.
 
 ### BUG FIXES
 *  Fixed SSCP issue in `summary.lm.rrpp` that produces matrices of 0s.  The issue was inherent to `lm.rrpp`, from erroneous assignment of models, only for output (not for mechanics).